feat: unifica headers 2 e 3
