+++ CHANGELOG.md
@@ -1,16 +1,20 @@
 # CHANGELOG
 
 ## US3.1 - PR 1
+
 ### Adicionado
+
 - Primeiro Layout
 - Primeiro Header
 - Estilizações iniciais
 
 ## US3.2 - PR 2
+
 ### Adicionado
+
 - Quatro Layouts
-- Dois Headers
+- Segundo Header
 - Util de Ícone
-- Componentes: NavBar, PerfilBox, TagBox e TagsPopulares
+- Componentes: NavBar, PerfilBox, TagBox, TagsPopulares e Barra de Pesquisa
 - Reset.css personalizado
 - Responsividade
