+++ src/components/layouts/Layout3.jsx
@@ -3,7 +3,7 @@ import NavBar from './NavBar';
 import TagsPopulares from './TagsPopulares';
 import tags from '../temporario/tags';
 
-//Layout Padrão com Header simples e Tags Populares
+//Layout Padrão com Header completo e Tags Populares
 //É possível definir o conteúdo que do lado esquerdo do Layout utilizando o parâmetro Conteudo
 //Atributos de Conteudo são definidos em conteudoProps
 function Layout({ Conteudo, conteudoProps }) {
